pyproject.toml
[tool.poetry]
name = "crypto"
version = "0.1.0"
description = ""
authors = ["Le Xuan Manh <lexuanmanh101199@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
mysqlclient = "^2.1.1"
drf-spectacular = "^0.24.2"
django-rest-marshmallow = "^4.0.2"
marshmallow = "^3.18.0"
djoser = "^2.1.0"
djangorestframework-simplejwt = "4.8.0"
zipp = "^3.10.0"
binance-futures-connector = "^3.2.0"
binance-connector = "^1.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
